requirements.txt
schedule==1.2.0
yfinance==0.2.29
pytz==2023.3.post1
PyYAML==6.0.1
tefas-crawler==0.3.4
pandas==2.1.1